BSP v0.9 beta
This is version 0.9 beta of my Node builder for DOOM.  Some credit should
go to Raphael Quinet, cos I used (re-wrote) some of the code from DEU.
There seem to still be a few problems (sometimes it will exit with a error
message (... Trouble in .... ) this is cos I have to improve my Node line
checker to pick more sensible lines. It only seems to do this on E3L2,
but I haven't tried all of the maps yet.), but I am working on them as
quickly as possible.

Currently it always outputs the finished PWAD file as TMP.WAD in the current
directory, cos I don't want to be responsible for trashing anyone's maps.

Also, if this program was continually run on a map, it would overload with
unused vertices, because the unused ones are not thrown away.

If you have any really groovy ways of picking the node lines, please let me
know, cos currently for E1L1 I am producing about 50% more nodes that the
original maps.